+++ src/types/object/test.rs
@@ -8,7 +8,7 @@ mod test_core {
 
     use crate::builder::ArcPointer;
     use crate::config::BucketBase;
-    use crate::types::object::invalid::InvalidObjectBaseKind;
+    use crate::types::object::base::invalid::InvalidObjectBaseKind;
     use crate::types::object::{FromOss, InvalidObjectDir};
     use crate::types::object::{InvalidObjectBase, InvalidObjectPath, ObjectBase};
     use crate::{BucketName, ObjectDir, ObjectPath};
